requirements.txt
Package                         Version
------------------------------- ---------
backports.cached-property       1.0.1
blessings                       1.7
bpython                         0.21
certifi                         2020.12.5
chardet                         4.0.0
curtsies                        0.3.5
cwcwidth                        0.1.4
greenlet                        1.0.0
grpcio                          1.37.0
grpcio-tools                    1.37.0
idna                            2.10
pip                             21.1.1
protobuf                        3.15.8
Pygments                        2.8.1
pyxdg                           0.27
requests                        2.25.1
robotframework                  4.0.1
robotframework-assertion-engine 0.0.6
robotframework-browser          4.3.0
robotframework-pythonlibcore    2.2.1
setuptools                      40.8.0
six                             1.15.0
typing-extensions               3.10.0.0
urllib3                         1.26.4
wrapt                           1.12.1
